SurfQuest_Backend/src/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-django ==4.7.0
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
gunicorn==20.1.0